+++ internal/trayapp/notify_other.go
@@ -2,7 +2,7 @@
 
 package trayapp
 
-func NotifySettingsChanged() error {
+func NotifySettingsChanged(configDir string) error {
 	// 非 Windows 平台当前没有托盘消息窗口，保存设置即可。
 	return nil
 }